REFINE: Replace StaticToken with GcsToken for improved token handling in AsyncStorage
