+++ client/amapy-plugin-gcs/amapy_plugin_gcs/transporter/async_gcs/async_storage.py
@@ -6,25 +6,7 @@ from typing import Any
 from aiofiles import open as file_open
 from gcloud.aio.storage import Storage
 
-API_ROOT = 'https://www.googleapis.com/storage/v1/b'
-
-
-class StaticToken:
-    """A minimal token wrapper for a pre-built GCP OAuth2 access token.
-
-    ``gcloud.aio.storage.Storage`` calls ``await token.get()`` to obtain
-    the bearer string. This class satisfies that interface without performing
-    any additional network requests.
-    """
-
-    def __init__(self, credentials: dict) -> None:
-        self._access_token = credentials.get("access_token")
-
-    async def get(self) -> str:
-        return self._access_token
-
-    async def close(self) -> None:  # called by Storage on session teardown
-        pass
+from amapy_plugin_gcs.gcs_token import GcsToken
 
 
 class AsyncStorage(Storage):
@@ -33,7 +15,7 @@ class AsyncStorage(Storage):
                  session=None) -> None:
         # check for access_token in the credentials
         if credentials.get("access_token"):
-            super().__init__(session=session, token=StaticToken(credentials))
+            super().__init__(session=session, token=GcsToken(credentials))
         else:
             super().__init__(session=session, service_file=io.StringIO(json.dumps(credentials)))
